"Refactor locking package to align with transaction processing module structure"
