+++ src/main/java/info/mackiewicz/bankapp/system/transaction/processing/locking/LockingConfig.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.system.locking;
+package info.mackiewicz.bankapp.system.transaction.processing.locking;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
