+++ src/main/java/info/mackiewicz/bankapp/system/transaction/processing/locking/LockingStrategy.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.system.locking;
+package info.mackiewicz.bankapp.system.transaction.processing.locking;
 
 import info.mackiewicz.bankapp.core.account.exception.AccountLockException;
