+++ src/main/java/info/mackiewicz/bankapp/system/transaction/processing/locking/LockingUtils.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.system.locking;
+package info.mackiewicz.bankapp.system.transaction.processing.locking;
 
 import java.util.Map;
 import java.util.Random;
