+++ src/test/java/info/mackiewicz/bankapp/system/locking/AccountLockingStrategyTest.java
@@ -1,6 +1,9 @@
 package info.mackiewicz.bankapp.system.locking;
 
 import info.mackiewicz.bankapp.core.account.exception.AccountLockException;
+import info.mackiewicz.bankapp.system.transaction.processing.locking.AccountLockingStrategy;
+import info.mackiewicz.bankapp.system.transaction.processing.locking.LockingConfig;
+import info.mackiewicz.bankapp.system.transaction.processing.locking.LockingUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
